results/run-20230924-033218_LLMatic-zero_shot_donnemartin-system-design-oop-whatsapp-donnemartin-system-design-oop-url_shortener-donnemartin-system-design-oop-dropbox-donnemartin-system-design-oop-twitter_0_5-runs_log/dropbox/1/LLMatic/cloudsafe/app/sharing_controller.py
from flask import Blueprint, request
from cloudsafe.app.sharing_service import SharingService

sharing = Blueprint('sharing', __name__)
sharing_service = SharingService()

@sharing.route('/generate_share_link', methods=['POST'])
def generate_share_link():
	data = request.get_json()
	response = sharing_service.generate_share_link(**data)
	return {'message': response}, 200

@sharing.route('/set_expiry_date', methods=['PUT'])
def set_expiry_date():
	data = request.get_json()
	response = sharing_service.set_expiry_date(**data)
	return {'message': response}, 200

@sharing.route('/set_password', methods=['PUT'])
def set_password():
	data = request.get_json()
	response = sharing_service.set_password(**data)
	return {'message': response}, 200

@sharing.route('/invite_user', methods=['POST'])
def invite_user():
	data = request.get_json()
	response = sharing_service.invite_user(**data)
	return {'message': response}, 200

@sharing.route('/set_permissions', methods=['PUT'])
def set_permissions():
	data = request.get_json()
	response = sharing_service.set_permissions(**data)
	return {'message': response}, 200
